Delete temporary dll if ejecting or injector is exiting

+++ parcel-thief/thief-injector/src/main.rs
@@ -90,7 +90,9 @@ async fn main() -> Result<(), anyhow::Error> {
                         println!("Failed to eject: {err:?}");
                     }
                 };
+            }
 
+            if should_eject || should_exit {
                 std::fs::remove_file(&copied_dll_path).context("Failed to delete temporary dll")?;
             }
